Refactor folder and file structure to enhance file management
