+++ src/common/slice.js
@@ -7,9 +7,9 @@ import {
   fetchRestaurant,
   postLogin,
   postReview,
-} from './services/api';
+} from '../services/api';
 
-import { saveItem } from './services/storage';
+import { saveItem } from '../services/storage';
 
 import { equal } from './utils';
