+++ src/containers/CategoriesContainer.jsx
@@ -5,9 +5,9 @@ import { useDispatch, useSelector } from 'react-redux';
 import {
   selectCategory,
   loadRestaurants,
-} from './slice';
+} from '../common/slice';
 
-import { get } from './utils';
+import { get } from '../common/utils';
 
 export default function CategoriesContainer() {
   const dispatch = useDispatch();
