+++ src/pages/LoginPage.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 
-import LoginFormContainer from './LoginFormContainer';
+import LoginFormContainer from '../containers/LoginFormContainer';
 
 export default function LoginPage() {
   return (
